+++ docs/CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 -----------------------
 - Enh: Added topic picker to edit form
 - Enh: Added "Add all space members to event" feature
+- Enh: Added calendar event title
+- Fixed: encoding issue in calendar view
+- Chg: Min HumHub version 1.3
 
 0.6.17  (July 4, 2018)
 -----------------------